+++ core/src/main/kotlin/sift/core/dsl/Parameters.kt
@@ -1,7 +1,6 @@
 package sift.core.dsl
 
 import sift.core.api.*
-import sift.core.element.AsmType
 import sift.core.element.MethodNode
 import sift.core.element.ParameterNode
 import sift.core.entity.Entity
@@ -33,8 +32,8 @@ class Parameters internal constructor(
     /**
      * Filters fields to include only those with a type matching type.
      */
-    fun filterType(type: AsmType) {
-        action += Action.Parameter.FilterType(type)
+    fun filterType(type: Type) {
+        action += Action.Parameter.FilterType(type.asmType)
     }
 
     override fun scope(